Compute List at Startup
I now compute the list of items once, when the server starts, rather than on
each request.
This does mean, though, that the app must be restarted in order to pick up
changes.
This could be done periodically, when no one is making requests, and it will be
fast for them when they do.

It's a trade I'm willing to make.